requirements.txt
# Core scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# JavaScript-rendered pages (Angular SPA: esoa.sbimf.com, corporate.sbimf.com)
playwright==1.43.0

# Tokenization (chunking)
tiktoken==0.7.0

# Embeddings (Phase 3)
openai==1.30.0
sentence-transformers==3.0.0   # fallback: all-MiniLM-L6-v2
chromadb==0.5.0

# Database
psycopg2-binary==2.9.9         # PostgreSQL (production)
# sqlite3 is part of Python stdlib — used for local dev without a Postgres server

# Utilities
python-dotenv==1.0.1
tqdm==4.66.4                   # progress bars for batch embedding
